requirements.txt
streamlit>=1.28.0
anthropic>=0.40.0
python-docx>=1.1.0
pypdf>=4.0.0
